refactor: migrate hooks
